Added explanation to "Methods" section
I broke up the code and added explanation between the breaks. I didn't touch anything before the "Methods" heading, but the code in the section underwent some small changes (changed seed from 2022 to 1, changed K value to 17) and repetitive code at the bottom of the document was deleted, as per Brian's request.